# contracts/openapi-diet-management-v1.yaml
openapi: 3.1.0
info:
  title: Diet Management API
  version: v1-fixed
  description: >
    스냅샷 계약 (SSOT 본문: docs/requirements/feature-diet-management-api-contract-v1.md).
    구현체는 본 파일과 문서를 함께 유지한다.
servers:
  - url: http://localhost:3000
    description: 로컬 스텁 (apps/server)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, details, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        traceId: { type: string }
...
    Dashboard:
      type: object
      properties:
        newUsers: { type: integer }
        activeUsers: { type: integer }
        mealRecordCount: { type: integer }
        inquiryCount: { type: integer }
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: includeInactive
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer, default: 15 }
      responses:
        '200': { description: OK }
        '403': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /admin/users/{id}/deactivate:
    patch:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/foods:
    get:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: includeInactive
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer, default: 15 }
      responses:
        '200': { description: OK }
    post:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Created }
  /admin/foods/{id}:
    put:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/foods/{id}/deactivate:
    patch:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/inquiries:
    get:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: includeInactive
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer, default: 15 }
      responses:
        '200': { description: OK }
  /admin/inquiries/{id}/status:
    patch:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/inquiries/{id}/deactivate:
    patch:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/notices:
    get:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: includeInactive
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer, default: 15 }
      responses:
        '200': { description: OK }
    post:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Created }
  /admin/notices/{id}:
    put:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/notices/{id}/deactivate:
    patch:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /admin/dashboard:
    get:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
  /admin/stats/reaggregate:
    post:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      responses:
        '202': { description: Accepted }